Removed commented lines.
Realised the lines:
.left-items {
  justify-content: flex-start;
}

.right-items {
  justify-content: flex-end;
}

don't do anything, and that "justify-content: space between"
which was written in .header does the job of separting
the two containers, thereby creating that space in the middle.

"align-items: center" which was added in .left-items,
.right-items... does the job of centralising the logo (irrespective
of font size,) and links.

alright. that's a lenghty commit message. Bye.