Modify to create an entire hoempage.
Fixed the header such that links aligned center at the top,
left-most and right-most, same for footer, except footer
remained fixed at the bottom.
Created containers for image and input box, then
positioned them in the center of the screen, alongside
the buttons right below them.

All of these come together to give the look of the
The Odin Project's hoempage.