Create the Holy Grail of layout.
Fixed the header and footer at the top and bottom
respectively, sing flex on the body tag.
Positioned the side nav bar on the left side using
flex on a div container that held both the container of the
sidebar and the container of the cards.
Centalised horizontally and vertically the text in
the footer.

Alright. Bye.